improve broker notifications (#564)
only send notifications to platforms that are active or that were active recently, i.e. no longer than the period for which we keep notifications for. If the platform was not connected for more than this period, the SM agent will anyways run full reconciliation so no notifications are needed.